+++ src/map.rs
@@ -5,7 +5,7 @@ use sdl2::pixels::Color;
 use sdl2::rect::Rect;
 use sdl2::render::{Texture, TextureCreator};
 use sdl2::surface::Surface;
-use sdl2::video::{Window, WindowContext};
+use sdl2::video::WindowContext;
 
 use crate::system::System;
 use crate::{MAP_CASE_SIZE, MAP_SIZE};
@@ -44,8 +44,8 @@ fn draw_in_map(
             None,
             surface_map,
             Rect::new(
-                pos_x as i32 * MAP_CASE_SIZE,
-                pos_y as i32 * MAP_CASE_SIZE,
+                pos_x as i32 * MAP_CASE_SIZE as i32,
+                pos_y as i32 * MAP_CASE_SIZE as i32,
                 surface.width(),
                 surface.height(),
             ),
@@ -56,8 +56,8 @@ fn draw_in_map(
 
 pub struct Map<'a> {
     pub data: Vec<u8>,
-    pub x: i32,
-    pub y: i32,
+    pub x: i64,
+    pub y: i64,
     pub texture: Texture<'a>,
 }
 
@@ -65,8 +65,8 @@ impl<'a> Map<'a> {
     pub fn new(
         texture_creator: &'a TextureCreator<WindowContext>,
         rng: &mut ChaCha8Rng,
-        x: i32,
-        y: i32,
+        x: i64,
+        y: i64,
     ) -> Map<'a> {
         let tree =
             Surface::from_file("resources/tree.png").expect("failed to load `resources/tree.png`");
@@ -115,17 +115,19 @@ impl<'a> Map<'a> {
         let x = system.x() - self.x;
         let y = system.y() - self.y;
         let (s_x, pos_x, width) = if x < 0 {
-            (0, x * -1, (system.width() + x) as u32)
-        } else if x + system.width() > MAP_SIZE as i32 * MAP_CASE_SIZE {
-            let sub = system.width() - (system.width() + x - MAP_SIZE as i32 * MAP_CASE_SIZE);
+            (0, x * -1, (system.width() as i64 + x) as u32)
+        } else if x + system.width() as i64 > MAP_SIZE as i64 * MAP_CASE_SIZE {
+            let sub = system.width() as i64
+                - (system.width() as i64 + x - MAP_SIZE as i64 * MAP_CASE_SIZE);
             (x, 0, sub as u32)
         } else {
             (x, 0, system.width() as u32)
         };
         let (s_y, pos_y, height) = if y < 0 {
-            (0, y * -1, (system.height() + y) as u32)
-        } else if y + system.height() > MAP_SIZE as i32 * MAP_CASE_SIZE {
-            let sub = system.height() - (system.height() + y - MAP_SIZE as i32 * MAP_CASE_SIZE);
+            (0, y * -1, (system.height() as i64 + y) as u32)
+        } else if y + system.height() as i64 > MAP_SIZE as i64 * MAP_CASE_SIZE {
+            let sub = system.height() as i64
+                - (system.height() as i64 + y - MAP_SIZE as i64 * MAP_CASE_SIZE);
             (y, 0, sub as u32)
         } else {
             (y, 0, system.height() as u32)
@@ -134,8 +136,8 @@ impl<'a> Map<'a> {
             .canvas
             .copy(
                 &self.texture,
-                Rect::new(s_x, s_y, width, height),
-                Rect::new(pos_x, pos_y, width, height),
+                Rect::new(s_x as i32, s_y as i32, width, height),
+                Rect::new(pos_x as i32, pos_y as i32, width, height),
             )
             .expect("copy map failed");
     }
